+++ gdk/gdk_atoms.h
@@ -47,7 +47,7 @@ typedef struct {
 	char name[IDLENGTH];
 	uint8_t storage;	/* stored as another type? */
 	bool linear;		/* atom can be ordered linearly */
-	uint16_t size;		/* fixed size of atom */
+	uint32_t size;		/* fixed size of atom */
 
 	/* automatically generated fields */
 	const void *atomNull;	/* global nil value */
@@ -450,7 +450,7 @@ strEq(const char *l, const char *r)
 
 __attribute__((__pure__))
 static inline size_t
-VarHeapVal(const void *b, BUN p, int w)
+VarHeapVal(const void *b, BUN p, uint32_t w)
 {
 	switch (w) {
 	case 1:
